ESLint Plugin: Rename avoid-unguarded-get-range-at to no-unguarded-get-range-at
