+++ packages/eslint-plugin/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### New Features
 
-- New Rule: [`@wordpress/avoid-unguarded-get-range-at`](https://github.com/WordPress/gutenberg/blob/master/packages/eslint-plugin/docs/rules/avoid-unguarded-get-range-at.md)
+- New Rule: [`@wordpress/no-unguarded-get-range-at`](https://github.com/WordPress/gutenberg/blob/master/packages/eslint-plugin/docs/rules/no-unguarded-get-range-at.md)
 
 ## 2.3.0 (2019-06-12)
