+++ packages/eslint-plugin/rules/no-unguarded-get-range-at.js
